+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.3.2] / 2025-03-05
 ### Features
+- Support localized `InstallLocation` path.
 ### Updates
 - Add `FinalPathNameExtension` to fix localized `InstallLocation`. (Fix: #13)
 - Update `GetProcesses` to use `FinalPathNameExtension` to fix localized `InstallLocation`.